Modificaciones para los test de integracion

+++ chatbot-acceso-datos/src/main/java/co/com/santander/chatbot/accesodatos/controller/input/UsuarioInput.java
@@ -4,11 +4,12 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.validation.constraints.NotEmpty;
+import java.math.BigInteger;
 
 @Builder
 @Data
 public class UsuarioInput {
-    private int telefono;
+    private BigInteger telefono;
     @NotEmpty(message = "Cola identificación es obligatorio")
     private String colaIdentificacion;
 }
